+++ app/pb_kits/playbook/pb_layout/layout.rb
@@ -19,11 +19,12 @@ module Playbook
                   default: "md"
       prop :transparent, type: Playbook::Props::Boolean, default: false
       prop :variant, type: Playbook::Props::Enum,
-                  values: %w[light dark gradient],
-                  default: "light"
+                     values: %w[light dark gradient],
+                     default: "light"
       prop :layout, type: Playbook::Props::Enum,
-            values: %w[sidebar collection kanban],
+            values: %w[sidebar collection collection_detail kanban content],
             default: "sidebar"
+
 
       prop :responsive, type: Playbook::Props::Boolean, default: false
 
@@ -33,6 +34,10 @@ module Playbook
             generate_classname("pb_layout_kit", layout)
           when "kanban"
             generate_classname("pb_layout_kit", layout, responsive_class)
+          when "collection_detail"
+            generate_classname("pb_layout_kit", layout )
+          when "content"
+            generate_classname("pb_layout_kit", layout)
 	        else
 	          first_class = generate_classname("pb_layout_kit_sidebar","size_#{size}", position, variant, transparent_class)
 	          [first_class, full_class, collapse_class].reject(&:empty?).join(" ")